Mejorado el almacenamiento local
